DivsibleSumPairs.java
Complete the divisibleSumPairs function in the editor below. It should return the integer count of pairs meeting the criteria.
divisibleSumPairs has the following parameter(s):
n: the integer length of array 
ar: an array of integers
k: the integer to divide the pair sum by

Print the number of (i, j)  pairs where i < j and  a[i] + a[j]  is evenly divisible by k.