MigratingBirdCount.java
Complete the migratoryBirds function in the editor below. It should return the lowest type number of the most frequently sighted bird.
migratoryBirds has the following parameter(s):
arr: an array of integers representing types of birds sighted

Print the type number of the most common bird; if two or more types of birds are equally common, choose the type with the smallest ID number.