ProgrammerDayTimeMachine.java
Marie invented a Time Machine and wants to test it by time-traveling to visit Russia on the Day of the Programmer (the 256th day of the year) during a year in the inclusive range from 1700 to 2700.
From 1700 to 1918, Russia's official calendar was the Julian calendar; since 1919 they used the Gregorian calendar system. The transition from the Julian to Gregorian calendar system occurred in 1918, when the next day after January was February 14th.

Complete the dayOfProgrammer function in the editor below. It should return a string representing the date of the  day of the year given. 256th dayOfProgrammer has the following parameter(s):
year: an integer

Print the full date of Day of the Programmer during year y in the format dd.mm.yyyy, where dd is the two-digit day, mm is the two-digit month, and yyyy is y.