PageTurnCount.java
Brie’s Drawing teacher asks her class to open their books to a page number. Brie can either start turning pages from the front of the book or from the back of the book. She always turns pages one at a time. When she opens the book, page 1 is always on the right side:

Complete the pageCount function in the editor below. It should return the minimum number of pages Brie must turn.
pageCount has the following parameter(s):
n: the number of pages in the book
p: the page number to turn to

Print an integer denoting the minimum number of pages Brie must turn to get to page p.